🔨 Refactoring: post/PostPreview 하위 컴포넌트 분리
