+++ src/components/PostPreview/PostPreview.tsx
@@ -1,74 +1,37 @@
 import type { Post } from '@/types';
-import { Avatar } from '../Avatar';
-import Icon from '../Icon';
+import PostHeader from './PostHeader';
 import {
   PreviewContainer,
   PostHeaderContainer,
-  PostInfoContainer,
-  PostContent,
-  UserNameContainer,
-  AvatarContainer,
-  PostDetailInfoContainer
+  PostContent
 } from './PostPreview.style';
 
 interface PostPreviewProps {
-  post: Pick<Post, 'image' | 'title' | 'author' | 'createdAt' >;
+  post: Pick<Post, 'image' | 'title' | 'author' | 'createdAt'>;
   totalLikes: number;
   totalComments: number;
   noneProfile: boolean;
 }
 
-const PostPreview = ({ post, totalLikes, totalComments, noneProfile = false }: PostPreviewProps) => {
-  const { image, title, author, createdAt } = post;
-  const iconDescription = [
-    { name: 'favorite', size: 12, total: totalLikes },
-    { name: 'chat', size: 12, total: totalComments }
-  ];
+const PostPreview = ({
+  post,
+  totalLikes,
+  totalComments,
+  noneProfile = false
+}: PostPreviewProps) => {
+  const { title } = post;
   const previewTitle = title.substring(0, 100);
   // todo: 명상 시간 데이터는 아직 못받아왔으므로 추후에 제외시켜놓은 거 추가하기
 
-  const PostDetailInfo = () => {
-    return (
-      <PostDetailInfoContainer>
-        {createdAt}
-        {iconDescription.map((iconInfo, index) => {
-          return (
-            <div key={index}>
-              <Icon
-                name={iconInfo.name}
-                size={iconInfo.size}
-                color={'greyLight'}
-              />
-              {iconInfo.total}
-            </div>
-          );
-        })}
-      </PostDetailInfoContainer>
-    );
-  };
-  const PostHeaderWithUser = () => {
-    return (
-      <>
-        <AvatarContainer>
-          <Avatar
-            alt={'유저 프로필'}
-            src={image}
-            size={25}
-          />
-        </AvatarContainer>
-        <PostInfoContainer>
-          <UserNameContainer>{author.fullName}</UserNameContainer>
-          <PostDetailInfo />
-        </PostInfoContainer>
-      </>
-    );
-  };
-
-  const PostHeader = noneProfile === true ? PostHeaderWithUser : PostDetailInfo;
   return (
     <PreviewContainer>
       <PostHeaderContainer>
-        <PostHeader />
+        <PostHeader
+          post={post}
+          totalLikes={totalLikes}
+          totalComments={totalComments}
+          noneProfile={noneProfile}
+        />
       </PostHeaderContainer>
       <PostContent>{previewTitle}...</PostContent>
     </PreviewContainer>
